[IA64] Check CPL for break 0x80001 case
This is a patch which intends for fixing ltrace problem.  When I
execute a command "ltrace ps" on dom0, it makes dom0 hung.  It comes
from "break 0x80001" is shared by ltrace(CPL=3) and hpsim(CPL=2).
To avoid this problem, I just add check CPL in ia64_break_fault code.

Signed-off-by: Atsushi SAKAI <sakaia@jp.fujitsu.com>