Xen and tools now require gcc 3.4+ on x86.  - gcc-3.2 cannot handle some multi-line assertions in the Xen  sources. Noone noticed.  - gcc-3.3 has problems with alignment constraints inside typedefs.
gcc 3.4.0 is now three years old so I hope that everyone has an
up-to-date compiler, or can obtain a more up-to-date package for their
distribution. If not we may need to fall back to supporting gcc-3.3.x
as well.

Also clean up the way we do version checks, using the power of awk.

Signed-off-by: Keir Fraser <keir@xensource.com>