xend: Fix SrvDomain.op_save.
The xen.xend.server.SrvDomain class provides an op_save method for
saving guests. This just calls into the domain_save method on
xen.xend.XenDomain The latter recently gained an extra 'checkpoint'
argument, but the SrvDomain class is not providing this. The attached
patch simply makes the 'checkpoint' arg default to 'False' if not
provided by the caller which makes old calling convention.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>